+++ src/main/java/org/folio/entlinks/domain/entity/AuthoritySourceType.java
@@ -1,0 +1,17 @@
+package org.folio.entlinks.domain.entity;
+
+public enum AuthoritySourceType {
+  FOLIO("folio"),
+
+  LOCAL("local");
+
+  private final String value;
+
+  AuthoritySourceType(String value) {
+    this.value = value;
+  }
+
+  public String getValue() {
+    return value;
+  }
+}